add client superclass
