+++ src/test/java/org/example/Client.java
@@ -1,0 +1,18 @@
+package org.example;
+
+import io.restassured.http.ContentType;
+import io.restassured.specification.RequestSpecification;
+
+import static io.restassured.RestAssured.given;
+
+public class Client {
+    protected static final String BASE_URI = "https://stellarburgers.nomoreparties.site";
+    private static final String BASE_PATH = "/api";
+
+    public RequestSpecification spec() {
+        return given().log().all()
+                .contentType(ContentType.JSON)
+                .baseUri(BASE_URI)
+                .basePath(BASE_PATH);
+    }
+}
