+++ src/test/java/org/example/orders/OrdersClient.java
@@ -1,0 +1,6 @@
+package org.example.orders;
+
+import org.example.Client;
+
+public class OrdersClient extends Client {
+}
