+++ src/test/java/org/example/user/UserCredentials.java
@@ -1,3 +1,5 @@
+package org.example.user;
+
 import lombok.Data;
 
 @Data
